+++ src/alpss/plotting/__init__.py
@@ -1,1 +1,2 @@
 from alpss.plotting.plots import plot_results, plot_voltage
+from alpss.plotting.hel import plot_hel_detection